fix: fail closed when session creation fails

+++ app/controllers/oidc_accounts_controller.rb
@@ -36,18 +36,30 @@ class OidcAccountsController < ApplicationController
     user = User.authenticate_by(email: params[:email], password: params[:password])
 
     if user&.active?
-      # Create the OIDC identity link
-      oidc_identity = OidcIdentity.create_from_omniauth(
-        build_auth_hash(@pending_auth),
-        user
-      )
+      linked = user.transaction do
+        OidcIdentity.create_from_omniauth(
+          build_auth_hash(@pending_auth),
+          user
+        )
 
-      # Log account linking
-      SsoAuditLog.log_link!(
-        user: user,
-        provider: @pending_auth["provider"],
-        request: request
-      )
+        SsoAuditLog.log_link!(
+          user: user,
+          provider: @pending_auth["provider"],
+          request: request
+        )
+
+        unless user.otp_required?
+          @session = create_session_for(user)
+          raise ActiveRecord::Rollback unless @session
+        end
+
+        true
+      end
+
+      unless linked
+        redirect_to new_session_path, alert: t("sessions.openid_connect.failed")
+        return
+      end
 
       # Clear pending auth from session
       session.delete(:pending_oidc_auth)
@@ -56,7 +68,6 @@ class OidcAccountsController < ApplicationController
         session[:mfa_user_id] = user.id
         redirect_to verify_mfa_path
       else
-        @session = create_session_for(user)
         notice = if accept_pending_invitation_for(user)
           t("invitations.accept_choice.joined_household")
         else
@@ -162,6 +173,9 @@ class OidcAccountsController < ApplicationController
           @user
         )
 
+        @session = create_session_for(@user)
+        raise ActiveRecord::Rollback unless @session
+
         true
       end
     rescue ActiveRecord::RecordInvalid, ActiveRecord::RecordNotUnique => e
@@ -184,7 +198,6 @@ class OidcAccountsController < ApplicationController
       # Clear pending auth from session
       session.delete(:pending_oidc_auth)
 
-      @session = create_session_for(@user)
       notice = if invitation.present?
         t("invitations.accept_choice.joined_household")
       elsif accept_pending_invitation_for(@user)
